# use faster matching