--- a/.travis.yml
+++ b/.travis.yml
@@ -1,13 +1,14 @@
 language: ruby
 rvm:
-  - 2.4.1
-  - 2.3.3
-  - 2.2.7
-  - 2.1.8
+  - 2.5.1
+  - 2.4.4
+  - 2.3.7
+  - 2.2.10
+  - 2.1.10
   - 2.0.0
   - 1.9.3
   - jruby-19mode
   - ruby-head
-  - jruby-9.1.8.0
+  - jruby-9.1.17.0
 before_install:
   - gem install bundler