--- a/.travis.yml
+++ b/.travis.yml
@@ -7,8 +7,9 @@ rvm:
   - 2.1.10
   - 2.0.0
   - 1.9.3
-  - jruby-19mode
   - ruby-head
+  - jruby-19mode
   - jruby-9.1.17.0
+  - jruby-9.2.0.0
 before_install:
   - gem install bundler