Faster matching:

 * accessing Regexp using `patterns[:regex]` is slightly faster than `patterns['regex']` (0.05 performance increase),
 * Ruby 2.4.0 introduced `match?` for faster regular expression matching, using `match?` to verify regular expression and then calling `match` to return matches is faster than using `match` for every iteration (0.32 performance increase).

```
Warming up --------------------------------------
               match   125.000  i/100ms
              match?   166.000  i/100ms
Calculating -------------------------------------
               match      1.266k (± 2.1%) i/s -      6.375k in   5.037348s
              match?      1.729k (± 2.1%) i/s -      8.798k in   5.090618s

Comparison:
              match?:     1729.1 i/s
               match:     1266.1 i/s - 1.37x  slower
```