@wallin 2018-09-06
Thanks for finding this @filiptepper. Could you perhaps make the method definition dynamically based on the version instead of having a conditional inside the method?