@filiptepper 2018-09-10
@wallin Something like https://github.com/ua-parser/uap-ruby/pull/36/commits/089083cde9d7703937c0765f55853b17d0eed7f9 will work?