@wallin 2018-09-11
Looks good @filiptepper. Should we also check for JRuby >9.2.0.0?