@filiptepper 2018-09-11
@wallin I added it, since it changes MRI Ruby compatibility to 2.5.